+++ spec/package_json/managers/yarn_classic_like_spec.rb
@@ -150,6 +150,63 @@ RSpec.describe PackageJson::Managers::YarnClassicLike do
       end
     end
 
+    it "supports adding dependencies with exact version" do
+      with_package_json_file do
+        result = manager.add(["example"], exact: true)
+
+        expect(result).to be(true)
+        expect_manager_to_be_invoked_with("add --exact example")
+        expect_package_json_with_content({
+          "dependencies" => {
+            "example" => "0.0.0"
+          }
+        })
+      end
+    end
+
+    it "supports adding dev dependencies with exact version" do
+      with_package_json_file do
+        result = manager.add(["example"], type: :dev, exact: true)
+
+        expect(result).to be(true)
+        expect_manager_to_be_invoked_with("add --dev --exact example")
+        expect_package_json_with_content({
+          "devDependencies" => {
+            "example" => "0.0.0"
+          }
+        })
+      end
+    end
+
+    it "supports adding multiple packages with exact version" do
+      with_package_json_file do
+        result = manager.add(%w[example example2], exact: true)
+
+        expect(result).to be(true)
+        expect_manager_to_be_invoked_with("add --exact example example2")
+        expect_package_json_with_content({
+          "dependencies" => {
+            "example" => "0.0.0",
+            "example2" => "0.0.0"
+          }
+        })
+      end
+    end
+
+    it "supports adding optional dependencies with exact version" do
+      with_package_json_file do
+        result = manager.add(["example"], type: :optional, exact: true)
+
+        expect(result).to be(true)
+        expect_manager_to_be_invoked_with("add --optional --exact example")
+        expect_package_json_with_content({
+          "optionalDependencies" => {
+            "example" => "0.0.0"
+          }
+        })
+      end
+    end
+
     context "when the group type is not supported" do
       it "raises an error" do
         expect { manager.add([], type: :unknown) }.to raise_error(PackageJson::Error)