+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+### Added
+
+- Add support for `exact` parameter in `add` method to install packages with
+  exact versions ([#29](https://github.com/shakacode/package_json/pull/29) by
+  [@justin808](https://github.com/justin808))
+
 ## [0.1.1] - 2025-11-04
 
 ### Changed